move utils config plots to tools
